float64

+++ audio/audio.py
@@ -76,11 +76,11 @@ class Audio:
 
         if file_path:  # If a file path is provided, use the original behavior
             os.makedirs(os.path.dirname(file_path), exist_ok=True)
-            wavfile.write(file_path, self.sample_rate, self.ndarray.astype(np.float32))
+            wavfile.write(file_path, self.sample_rate, self.ndarray.astype(np.float64))
         else:
             # Use a BytesIO buffer instead of a file
             buffer = io.BytesIO()
-            wavfile.write(buffer, self.sample_rate, self.ndarray.astype(np.float32))
+            wavfile.write(buffer, self.sample_rate, self.ndarray.astype(np.float64))
             buffer.seek(0)  # Move to the start of the buffer for reading if needed
             
             # You can return or use the buffer for further processing
